Ajout de typeMonument et ajout de methode dans VarianteCarteRepository et AncienNomRepository

+++ src/main/java/fr/m2miage/geocartebck/dao/CartePostaleRepository.java
@@ -3,6 +3,6 @@ package fr.m2miage.geocartebck.dao;
 import fr.m2miage.geocartebck.model.CartePostale;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartePostaleRepository extends JpaRepository<CartePostale, Integer>
+public interface CartePostaleRepository extends JpaRepository<CartePostale, Long>
 {
 }
